requirements.txt
beautifulsoup4
requests
pwntools
PyCryptoDome
pycrypto